Return the munged config if it has already been munged
When using the PGHero gem, which tries to establish its own connection,
the `#munge_config` monkeypatch gets called again. For some reason,
however, this time through the keys in the hash are symbols instead of
strings, so the `config.fetch('adapter')` fails.

We probably do not want to re-munge the config again anyways (so that
both 'adapter' and 'real_adapter' would == 'honeycomb'), so just return
the config if we detect that its already been munged.